style: format with forge fmt
style: format with forge fmt

+++ src/levels/02-Fallout/Fallout.sol
@@ -5,7 +5,8 @@ import "openzeppelin-contracts/math/SafeMath.sol";
 
 contract Fallout {
     using SafeMath for uint256;
-    mapping(address => uint) allocations;
+
+    mapping(address => uint256) allocations;
     address payable public owner;
 
     // @audit-issue typo in the constructor name
@@ -43,7 +44,7 @@ contract Fallout {
         msg.sender.transfer(address(this).balance);
     }
 
-    function allocatorBalance(address allocator) public view returns (uint) {
+    function allocatorBalance(address allocator) public view returns (uint256) {
         return allocations[allocator];
     }
 }
